requirements.txt
# PSI Lab Metrics Tool - 依存関係
#
# 基本要件: Python 3.8+
#
# インストール方法:
# pip install -r requirements.txt

# HTTP クライアント（PSI API通信）
requests>=2.28.0

# YAML設定ファイル処理
PyYAML>=6.0.0

# データ処理・CSV操作（大量データ対応）
pandas>=1.5.0

# CLI インターフェース
click>=8.1.0

# 環境変数管理
python-dotenv>=0.19.0

# URL解析・検証
urllib3>=1.26.0

# 型ヒント（Python 3.8対応）
typing-extensions>=4.0.0; python_version < "3.10"

# 開発・テスト用依存関係（オプション）
# 以下は開発時のみ必要
# pytest>=7.0.0
# pytest-cov>=4.0.0
# flake8>=5.0.0
# mypy>=0.991
